backend/requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
pydantic~=2.11.2
langchain~=0.3.23
python-dotenv~=1.1.0
openai-agents
langchain-openai~=0.3.12